#147-refactor: adjust my events tab
