+++ frontend/src/components/organisms/tabs/myEvents.tsx
@@ -1,0 +1,33 @@
+import { useQuery } from "react-query";
+import { Typography } from "@mui/material";
+
+import { EventUs } from "types";
+import { userApi } from "api";
+
+import { Loader } from "components/atoms";
+import { EventCard } from "components/molecules";
+
+const MyEvents = () => {
+  const { isLoading, data: events } = useQuery("events", () =>
+    userApi
+      .getEventsByOrganizer()
+      .then((response) => response.data as EventUs[])
+  );
+
+  return (
+    <section>
+      <Typography variant="h4">Mis Eventos</Typography>
+      {isLoading ? (
+        <Loader />
+      ) : (
+        <section className="mt-6 grid w-full grid-cols-1 gap-5 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 2xl:grid-cols-5">
+          {events?.map((event) => (
+            <EventCard event={event} />
+          ))}
+        </section>
+      )}
+    </section>
+  );
+};
+
+export default MyEvents;
