+++ frontend/src/pages/profile/profile.tsx
@@ -5,8 +5,11 @@ import { Box, Tab, Tabs } from "@mui/material";
 import { Event, Info, LocationCity, Receipt } from "@mui/icons-material";
 
 import { TabPanel } from "components/molecules";
-import MyEvents from "./events";
-import { ProfileInfoTab, MyLocationsTab } from "components/organisms";
+import {
+  ProfileInfoTab,
+  MyLocationsTab,
+  MyEventsTab,
+} from "components/organisms";
 import Page from "../page";
 
 const tabs = {
@@ -44,7 +47,7 @@ const ProfilePage = () => {
         </Tabs>
       </Box>
       <TabPanel value={activeTabIndex} index={0}>
-        <MyEvents/>
+        <MyEventsTab />
       </TabPanel>
       <TabPanel value={activeTabIndex} index={1}>
         <MyLocationsTab />
